==> python_notebooks/4_Identification.ipynb <==
import tqdm
import pandas as pd

import matplotlib.pyplot as plt
import numpy as np
from sklearn import datasets, linear_model
from sklearn.metrics import mean_squared_error, r2_score

# Read the db file and clustering results
db_file = pd.read_csv("../susdat_2022-08-19-155358.csv",sep="\t")
feature_file_name = "analysis_results/coverage0.9_correlation0.6.csv"
clustered_data_file = pd.read_csv(feature_file_name)

db_file.columns

def calculate_ppm_error_between_two_masses(exact_mass,observed_mass): # Calculate ppm error between a two masses
    mass_difference = observed_mass-exact_mass
    ppm_error = (mass_difference * 1000000) / exact_mass
    return (ppm_error)

all_masses = db_file["M-H-"].to_list()
log_k = db_file["logKow_EPISuite"].to_list()

ppm_error_cutoff = 5
count = 0
output_results = []
# Compare the M-H- value to the m/z and get the hits and consoliate the results
for index,row in tqdm.tqdm(clustered_data_file.iterrows(),leave=True,position=0,total=len(clustered_data_file)):
    if row["Met_ID"]:
        mz_cf = row["mz_cf"]
        hits = [i for i,mass in enumerate(all_masses) if abs(calculate_ppm_error_between_two_masses(mz_cf,mass)) <= ppm_error_cutoff]
        names = ""
        for indexes in hits:
            names += db_file.iloc[indexes]["Name"] + " % "
        if len(hits) > 0:
            count += 1
            met_id_found_names = {}
            met_id_found_names["met_id"] = int(row["Met_ID"])
            met_id_found_names["mz_cf"] = mz_cf
            met_id_found_names["rt_cf"] = row["rt_cf"]
            met_id_found_names["label"] = row["label"]
            met_id_found_names["intensity_cf"] = row["intensity_cf"]
            
            log_k_list = ""
            for indexes in hits:
                log_k_list += str(log_k[indexes]) + " % "
            met_id_found_names["log_k"] = log_k_list
            
            if row["real_name"] == "nan":
                met_id_found_names["real_name"] = "NA"
            else:
                met_id_found_names["real_name"] = row["real_name"]
            met_id_found_names["names"] = str(names)
            output_results.append(met_id_found_names)

# Read native compound files and fetch necessary values to make a lm model
native_cmpds = pd.read_csv("all_no_Br_Cl.csv")
native_cmpds = native_cmpds[native_cmpds["Native/IS"] == "Native"]
names = native_cmpds["Abbreviation"].to_list()
rt_values =  [(entry) for entry in native_cmpds["Rt(s)"].to_list()]
mz_values =  [(entry) for entry in native_cmpds["MIM"].to_list()]
log_k_list = [(entry) for entry in native_cmpds["Unnamed: 11"].to_list()]

# make rt mz feature
rt_mz_values = [[rt,mz] for rt,mz in zip(rt_values,mz_values)]

rt_list = np.array(rt_values)
rt_list = rt_list.reshape(-1,1)
rt_list = rt_mz_values
# Create linear regression object
regr = linear_model.LinearRegression()
regr.fit(rt_list, log_k_list)
y_pred = regr.predict(rt_list)

# The coefficients
print("Coefficients: ", regr.coef_)
# The mean squared error
print("Mean squared error: %.2f" % mean_squared_error(log_k_list, y_pred))
# The coefficient of determination: 1 is perfect prediction
print("Coefficient of determination: %.2f" % r2_score(log_k_list, y_pred))

# Making a df from the output_results
# File containing met_id	mz_cf	rt_cf	label	intensity_cf	log_k	real_name	names
out_df = pd.DataFrame.from_dict(output_results)
out_df

# Predicting for each rt,mz values
for index,rows in out_df.iterrows():
    mz = rows["mz_cf"]
    rt = rows["rt_cf"]
    pred_log_k = regr.predict([[rt,mz]])[0]
    condition = (out_df["mz_cf"] == mz) & (out_df["rt_cf"] == rt)
    out_df.loc[condition, 'pred_log_k_value'] = round(pred_log_k,2)

# Writing output to a file
out_df = out_df[['met_id', 'mz_cf', 'rt_cf', 'label', 'intensity_cf', 'log_k','pred_log_k_value', 'real_name', 'names']]
out_df.to_csv(feature_file_name[:-4] + "_identification.tsv",sep="\t")
